+++ CHANGES.md
@@ -1,3 +1,11 @@
+0.5.0
+-----
+* Rescheduler::Worker support routines and binary
+* New independent rk_args key setup to avoid contentions.
+* Service routine "weak" mutual exclusion to avoid contentions.
+* Lots of improvements for heavy queue operations.
+* Contention free stats
+
 0.4.0
 -----
 * MultiJson dependency (instead of JSON)